Animations / Upgrade
- Added animations to battering ram and door. Implemented their activations.
- Added particle effects to hitting door.
- Added upgrade canvas, implemented their scripts.
- Implemented DataManager all necessary upgrades.
- Both battering ram and Ai updates itself when an upgrade triggers.
- Added win animation to Ai that triggers when the door is broken.
- Ai stamina is reduced by half in each door hit.

+++ Castle Invasion/Assets/_ZestGames/Scripts/Ui/Hud.cs
@@ -27,9 +27,9 @@ namespace ZestGames
         }
 
         private void UpdateLevelText(int level) => levelText.text = $"Level {level}";
-        private void UpdateMoneyText(int money)
+        private void UpdateMoneyText(float money)
         {
-            collectableText.text = money.ToString();
+            collectableText.text = ((int)money).ToString();
             DOTweenUtils.ShakeTransform(transform, 0.25f);
         }
     }
